requirements.txt
Django>=3.2,<4.0
psycopg2-binary>=2.9,<3.0
djangorestframework
PyMuPDF
openai
transformers
torch
faiss-cpu
numpy